refactor: use i64/int8 for quantities where possible to reduce casting

+++ migrations/2025-02-02-214455_unit_costs/up.sql
@@ -4,7 +4,7 @@ CREATE TABLE unit_cost
     id         UUID          NOT NULL DEFAULT uuidv7(),
     unit_id    UUID          NOT NULL,
     resource   resource_type NOT NULL,
-    amount     INTEGER       NOT NULL DEFAULT 0,
+    amount     BIGINT        NOT NULL DEFAULT 0,
     created_at TIMESTAMPTZ   NOT NULL DEFAULT now(),
     updated_at TIMESTAMPTZ   NOT NULL DEFAULT now(),
